requirements.txt
asgiref==3.8.1
async-timeout==5.0.0
backports.zoneinfo==0.2.1
crispy-bootstrap5==2024.10
Django==4.2.16
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-redis==5.4.0
djangorestframework==3.15.2
pillow==10.4.0
PyJWT==2.9.0
PyMySQL==1.1.1
redis==5.2.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
